Adds build scripts
In order to automate organizing placement different binaries for
different platforms, this commit adds build scripts that will create
directories for different platforms and optimization levels for placing
the binary into. In addition, this commit adds the build shell script
for actually copying the binary into its desired directory.
This will be useful for organizing the different builds and the files
that are creating upon running the program. The alternative would be
manually making a directory every time and hand moving the binary.