Adds a basic image loading implementation and fixes logging
This commit adds a base implementation of image loading on the renderer.
The image loading handles any alpha blending that may be included in
images.

In addition, this commit adds a fix to logging. The log4rs crate always
has issues compiling or running on windows binaries. As a result, this
commit will be changing to env_logger as its primary logging crate.