Fixes logging for windows builds and removes app_config
This commit fixes an issue where log files couldn't be created on
windows builds, causing a panic. The timestamp used for naming a log
file contained illegal characters, in this case a colon `:`. To fix
that, any `:` is replaced with `-` in the timestamp.

In addition, this commit removes the now unused app_config. Choosing
instead to add environment variables in the build file depending on the
build profile being used.