Adds a main menu, and the structure of a menu
In order to add menus to the game, this commit implements the Menu
struct. This struct will contain a list of options which are build from
an object that implements the MenuOptionData trait. This trait will
force an option to have the name of an asset it's tied to, the name
of itself, and how to convert itself into the MenuOption Struct.
The point of this trait is to allow for an enum that lists all the
possible options of a menu, rather than it being in string form across
multiple locations.

In addition, this commit adds controls for the player. The controls have
been restricted to set actions to separate the code for actually
checking the keys that were pushed, and better organize the actions with
defined names. Both the game and menu states have their own separate
actions to help with this, even if some of the actions are similar
between the two.