Updates logging settings to rely on environment variables on runtime
To have a more idiomatic way of changing logging settings such as the
format and level, this commit changes the use of the `env!` macro to the
`std::env::var()` function to check for run time environment variables
rather than storing them at compile time.

In addition, this commit removes the personal bin directory and any
uneccessary code from the build script.