Changes the implementation for defining a menu.
In order to cut down on the amount of code needed to create a new menu,
this commit adds the `define_menu_items` macro. This macro will
allow for a very clear way of defining each menu item. The data for
creating each item will be used to implement the `MenuItemData` trait,
allowing for each menu item defined to be converted into the `MenuItem`
object. With each item in the menu knowning its `MenuItem` data, the
last method `full_list` will return the entire list of items defined
converted into `MenuItems`. This allows for an easy way to create a
menu from an enum defined in this `define_menu_items` macro.