Fixes text boxes not being rendered properly
This commit fixes a bug where each character in a text box would be
rendered as a single pixel.
This was caused by the fact that the position for each pixel was being
calculated incorrectly. Instead of using the x y of the internal
rectangle that is each character, the x y position of the top left for
each character was being used instead.

In addition, the top left was being incorrectly calculated. This
has been fixed by adjusting the top left for each character with its
given offset determined by the fontdue crate.