Changes how assets are handled and adds line method to renderer
This commit changes how assets are handled from an initialized struct to
static values of OnceLock<T>. This will allow for better access of these
assets without worrying about passing around an `Asset` object all over
the place.
This commit also moves all font handling to the `TextBox` struct and
`Asset` namespace. This improves handling fonts by not requiring the
`Renderer` to be passed all over the place.

In addition, this commit adds the `line` method to the renderer for
rastorizing vectors using the Bresenham algorithm implemented by the
[`Bresenham`](https://crates.io/crates/bresenham) crate.